+++ public/src/components/NoCards.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import {Text, View, Image, Dimensions, Button} from 'react-native';
 
-import styles from './css/card.js'
+import styles from '../style/card.js'
 
 class NoCards extends React.Component {
   render() {
