+++ public/src/img/check-mark.png
